pyproject.toml
[project]
name = "fastapitrainingproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "mypy (>=1.20.0,<2.0.0)",
    "ruff (>=0.15.9,<0.16.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
